003 Code/requirements.txt
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
filelock==3.9.0
fonttools==4.38.0
gdown==4.6.0
huggingface-hub==0.11.1
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.3
numpy==1.24.1
packaging==23.0
pandas==1.5.2
Pillow==9.4.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
scikit-learn==1.2.0
scipy==1.10.0
sentencepiece==0.1.97
simplejson==3.18.1
six==1.16.0
soupsieve==2.3.2.post1
threadpoolctl==3.1.0
timm==0.4.12
tokenizers==0.13.2
tqdm==4.64.1
transformers==4.25.1
typing-extensions==4.4.0
urllib3==1.26.14
WebOb==1.4.1
